Make a new component called RemoteNodeSelector
1

Update RemoteNodeSelector.tsx
